[INITIAL SOLUTION] can build gatewayd with docker
This solution provides a gatewayd instance in two docker containers, one for the database and the other for the app.
Be careful not to stop the database container accidentally, as this could cause the data to be lost.

Possible refactors, create another docker image that houses these images behind an nginx server.